requirements.txt
docker-py
boto
